+++ spring-security/src/main/java/com/self/security/utils/SecurityUtils.java
@@ -21,7 +21,7 @@ public class SecurityUtils {
         }
     }
 
-    public static Integer getUserId(){
+    public static Long getUserId(){
         return getAuthUser().getUserId();
     }
